Fixed lint errors (#35)
* Fixed lint errors

* Fixed another lint errors

* Update server.go and order.go

* Update route.go

* Formatter fixes

* Update route.go and server.go

* Update main.go, auth.pb.go, and 18 more files...

* Update go.mod and go.sum

* Update main.go, init.go, and 19 more files...

+++ order/internal/service/helper.go
@@ -2,6 +2,7 @@ package service
 
 import (
 	"context"
+
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 
@@ -46,7 +47,7 @@ func GetUserType(ctx context.Context, userID uint32) (string, error) {
 	userClient := userpb.NewUserServiceClient(conn)
 
 	req := &userpb.TypeRequest{
-		UserId: uint32(userID),
+		UserId: userID,
 	}
 
 	resp, err := userClient.GetType(ctx, req)
